:truck: Moved `NotificationHelper`
