+++ core/src/main/java/com/arnyminerz/escalaralcoiaicomtat/core/worker/MarkCompletedWorker.kt
@@ -18,7 +18,7 @@ import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.path.completion.reque
 import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.path.completion.request.MarkingDataInt
 import com.arnyminerz.escalaralcoiaicomtat.core.notification.ALERT_CHANNEL_ID
 import com.arnyminerz.escalaralcoiaicomtat.core.utils.toWorkData
-import com.arnyminerz.escalaralcoiaicomtat.notification.Notification
+import com.arnyminerz.escalaralcoiaicomtat.core.notification.Notification
 import com.google.firebase.firestore.FieldValue
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
